pyproject.toml
[project]
name = "gybs-tt"
version = "0.1.0"
description = ""
authors = [
    {name = "Constantine Zagorodnyi",email = "constantine2903@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "flask-restful (>=0.3.10,<0.4.0)",
    "flask-marshmallow (>=1.3.0,<2.0.0)",
    "marshmallow-sqlalchemy (>=1.4.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.5"
flake8 = "^7.1.2"
pytest = "^8.3.5"

